JS working properly
able to generate a random pokemon from first to current generation.
Will print the front sprite of that pokemon

FOR FUTURE
- Be able to create a team of 6 randomly generated pokemon
- Able to generate stats and items as well
- create a mini game where two people can generate 6 pokemon and be able to use them as cards which functions like a mini rock paper scissors...
      -> works as person 1 and 2 choose 1 out of their 6 randomly generated pokemon which will battle with each other resulting in a win, loss or tie.

there will be a total of 6 rounds and the winner of the round will be determined by who has a winning type advantage...

for example
water pokemon beats fire because water to fire is 2x weakness where as fire to water is 0.5x weakness.

a tie case scenario would be two dragon types as they are both 2x weakness to each other